+++ packages/common-ts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @eth-optimism/common-ts
 
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [17962ca9]
+  - @eth-optimism/core-utils@0.8.6
+
 ## 0.2.8
 
 ### Patch Changes